refactoring controllers package to internal/controller package as part of go/v4 prep changes in operator sdk v1.28
